Backend/requirements.txt
flask
flask-cors
transformers
torch
python-docx
PyMuPDF
python-dotenv
